// src/app/shared/models/user.ts
export interface User {
    firstName: string;
    lastName: string;
    phoneNumber: string;
    username: string;
    photoUrl: string;
    address: string;
    townCity: string;
    state: string;
    country: string;
    postalCodeZip: number;
    differentAddress: boolean;
    orderNote: string;
    provider: string;
    deliveryDate: string;
}